move defaultBackend back into backend.go
